Update README.md
The Resistance is a party game of social deduction. It is designed for five to ten players, lasts about 30 minutes, and has no player elimination. The Resistance is inspired by Mafia/Werewolf, yet it is unique in its core mechanics, which increase the resources for informed decisions, intensify player interaction, and eliminate player elimination.
Players are either Resistance Operatives or Imperial Spies. For three to five rounds, they must depend on each other to carry out missions against the Empire. At the same time, they must try to deduce the other players’ identities and gain their trust. Each round begins with discussion. When ready, the Leader entrusts sets of Plans to a certain number of players (possibly including himself/herself). Everyone votes on whether or not to approve the assignment. Once an assignment passes, the chosen players secretly decide to Support or Sabotage the mission. Based on the results, the mission succeeds (Resistance win) or fails (Empire win). When a team wins three missions, they have won the game.

This is a mod for Tabletop Simulator, meant to automate many aspects of the game and add new digital only roles.

Feature List:
* Assigns Roles and Deals Cards
* Reveals Teammates or Spies
* Deals Vote Tokens
* Calulate Votes
* Deals Mission Cards
* Analyzes Mission Result
* Tracks Winner
* Custom Roles